Reorganizes the js modules
* noves each class into its own files and let alephbet.js import them all
* also splits the jest tests for them
* rename AlephbetAdapter to TrackingAdapter to give it a more context
  specifig name
* remove lame adapter and its tests since it is basicall the same as the gimel adapter
* removes all build files from dist except the two .min files
* added a  `prepare` script which creates them when creating a package
* the `commonjs` files are now build into its own folder
* commonjs and esm build files are now git ignored and because of this I
  added .npmignore which doesn't ignore them wo they get packaged
