+++ src/adapters/local_storage_adapter.js
@@ -1,0 +1,14 @@
+import Storage from "../storage"
+
+const LocalStorageAdapter = {
+  namespace: "alephbet",
+
+  set(key, value) {
+    return new Storage(this.namespace).set(key, value)
+  },
+
+  get(key) {
+    return new Storage(this.namespace).get(key)
+  }
+}
+export default LocalStorageAdapter
